+++ tests/python/openassetio/pluginSystem/resources/pathA/modulePlugin.py
@@ -1,16 +1,18 @@
 """
-Provides a test PluginSystemPlugin implemented within a single file
+Provides a test PythonPluginSystemPlugin implemented within a single file
 module.
 """
 
-from openassetio.pluginSystem import PluginSystemPlugin
+from openassetio.pluginSystem import PythonPluginSystemPlugin
 
-class ModulePlugin(PluginSystemPlugin):
+
+class ModulePlugin(PythonPluginSystemPlugin):
     # pylint: disable=missing-class-docstring
 
     @classmethod
     def identifier(cls):
         return "org.openassetio.test.pluginSystem.resources.modulePlugin"
+
 
 # pylint: disable=invalid-name
 plugin = ModulePlugin